move some things into Core
