+++ src/slskd/Common/Compute.cs
@@ -44,22 +44,5 @@ namespace slskd
             using var sha1 = SHA1.Create();
             return BitConverter.ToString(sha1.ComputeHash(Encoding.UTF8.GetBytes(str))).Replace("-", string.Empty);
         }
-
-        public static string MaskHash(string str)
-        {
-            var hash = Sha1Hash(str);
-            hash = Convert.ToBase64String(Encoding.UTF8.GetBytes(hash));
-            hash = Regex.Replace(hash, @"[\d=]", string.Empty).ToLowerInvariant();
-
-            // in some very unlucky circumstances, the sha1 might end up being all or mostly numbers.
-            // if this is the case the resulting hash could be fewer than the 5 characters we need,
-            // so copy it until we get to 5
-            while (hash.Length < 5)
-            {
-                hash += hash;
-            }
-
-            return $"@@{hash.Substring(0, 5)}";
-        }
     }
 }
