+++ src/slskd/Core/QueueStrategy.cs
